requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
cssselect2==0.7.0
distlib==0.3.8
Django==5.0.2
django-decouple==2.1
django-filter==24.1
django-khalti==1.0
django-password-validators==1.7.3
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-standardized-errors==0.13.0
filelock==3.14.0
html5lib==1.1
identify==2.5.36
idna==3.6
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.6
lxml==5.2.1
nodeenv==1.8.0
oscrypto==1.3.0
pillow==10.2.0
platformdirs==4.2.2
pre-commit==3.7.1
prompt-toolkit==3.0.43
psycopg==3.1.16
pycparser==2.22
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.3
referencing==0.35.1
reportlab==4.0.9
requests==2.31.0
rest-framework-simplejwt==0.0.2
rpds-py==0.18.1
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.2
wcwidth==0.2.13
webencodings==0.5.1
